transfer-contract: Split phoenix and moonlight tests
